+++ src/main/java/com/avijit/together/server/service/CredentialService.java
@@ -20,6 +20,9 @@ import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
 
+import com.avijit.together.server.exception.ErrorCode;
+import com.avijit.together.server.exception.IErrorDetails;
+import com.avijit.together.server.exception.TogetherException;
 import com.avijit.together.server.model.Credential;
 import com.avijit.together.server.model.bean.CredentialBeanFactory;
 import com.avijit.together.server.repository.ICredentialRepository;
@@ -40,8 +43,7 @@ public class CredentialService implements IAuthenticationService, UserDetailsSer
 	/*
 	 * (non-Javadoc)
 	 * 
-	 * @see
-	 * com.avijit.together.server.service.IAuthenticationService#findAll(org.
+	 * @see com.avijit.together.server.service.IAuthenticationService#findAll(org.
 	 * springframework.data.domain.Pageable)
 	 */
 	@Override
@@ -52,13 +54,19 @@ public class CredentialService implements IAuthenticationService, UserDetailsSer
 	/*
 	 * (non-Javadoc)
 	 * 
-	 * @see
-	 * com.avijit.together.server.service.IAuthenticationService#findById(java.
+	 * @see com.avijit.together.server.service.IAuthenticationService#findById(java.
 	 * lang.String)
 	 */
 	@Override
-	public Credential findById(String authenticationId) {
-		return iCredentialRepository.findOne(UUID.fromString(authenticationId));
+	public Credential findById(String authenticationId) throws TogetherException {
+		try {
+			return iCredentialRepository.findOne(UUID.fromString(authenticationId));
+		} catch (IllegalArgumentException illegalArgumentException) {
+			throw new TogetherException(ErrorCode.INVALID_AUTHENTICATION_ID, IErrorDetails.INVALID_AUTHENTICATION_ID);
+		} catch (Exception exception) {
+			throw new TogetherException(ErrorCode.INVALID_AUTHENTICATION_ID,
+					String.format(IErrorDetails.INVALID_AUTHENTICATION_ID, authenticationId));
+		}
 	}
 
 	/*
@@ -69,34 +77,37 @@ public class CredentialService implements IAuthenticationService, UserDetailsSer
 	 * .together.server.model.Credential)
 	 */
 	@Override
-	public Credential save(Credential credential) {
-		credential.setId(UUID.randomUUID());
-		credential.setLastPasswordResetDate(new Date());
-		credential.setEnabled(true);
-		PasswordEncoder passwordEncoder = new BCryptPasswordEncoder();
-		String hashedPassword = passwordEncoder.encode(credential.getPassword());
-		credential.setPassword(hashedPassword);
+	public Credential save(Credential credential) throws TogetherException {
 		try {
+			credential.setId(UUID.randomUUID());
+			credential.setLastPasswordResetDate(new Date());
+			credential.setEnabled(true);
+			PasswordEncoder passwordEncoder = new BCryptPasswordEncoder();
+			String hashedPassword = passwordEncoder.encode(credential.getPassword());
+			credential.setPassword(hashedPassword);
 			return iCredentialRepository.save(credential);
 		} catch (Exception exception) {
-			return null;
+			throw new TogetherException(ErrorCode.AUTHENTICATION_NOT_ADDED, IErrorDetails.UNABLE_TO_ADD_AUTHENTICATION);
 		}
 	}
 
 	/*
 	 * (non-Javadoc)
 	 * 
-	 * @see
-	 * com.avijit.together.server.service.IAuthenticationService#delete(java.
+	 * @see com.avijit.together.server.service.IAuthenticationService#delete(java.
 	 * lang.String)
 	 */
 	@Override
-	public boolean delete(String authenticationId) {
+	public boolean delete(String authenticationId) throws TogetherException {
 		try {
-			iCredentialRepository.delete(UUID.fromString(authenticationId));
-			return true;
+			if (isExists(authenticationId)) {
+				iCredentialRepository.delete(UUID.fromString(authenticationId));
+				return true;
+			} else {
+				return false;
+			}
 		} catch (Exception exception) {
-			return false;
+			throw new TogetherException(ErrorCode.INVALID_AUTHENTICATION_ID, IErrorDetails.INVALID_AUTHENTICATION_ID);
 		}
 	}
 
@@ -108,11 +119,32 @@ public class CredentialService implements IAuthenticationService, UserDetailsSer
 	 */
 	@Override
 	public UserDetails loadUserByUsername(String username) throws UsernameNotFoundException {
-		Credential credential = iCredentialRepository.findByUsername(username);
-		if (credential == null) {
-			throw new UsernameNotFoundException(String.format("No user found with username '%s'.", username));
-		} else {
-			return CredentialBeanFactory.create(credential);
+		try {
+			Credential credential = iCredentialRepository.findByUsername(username);
+			if (credential == null) {
+				throw new UsernameNotFoundException(String.format(IErrorDetails.USER_USERNAME_NOT_FOUND, username));
+			} else {
+				return CredentialBeanFactory.create(credential);
+			}
+		} catch (Exception exception) {
+			throw new UsernameNotFoundException(String.format(IErrorDetails.USER_USERNAME_NOT_FOUND, username));
+		}
+	}
+
+	/*
+	 * (non-Javadoc)
+	 * 
+	 * @see
+	 * com.avijit.together.server.service.IAuthenticationService#isExists(java.lang.
+	 * String)
+	 */
+	@Override
+	public boolean isExists(String authenticationId) throws TogetherException {
+		try {
+			return iCredentialRepository.exists(UUID.fromString(authenticationId));
+		} catch (IllegalArgumentException illegalArgumentException) {
+			throw new TogetherException(ErrorCode.INVALID_AUTHENTICATION_ID,
+					String.format(IErrorDetails.INVALID_AUTHENTICATION_ID, authenticationId));
 		}
 	}
